+++ NotoriousTest.UnitTests/MultipleInfrastructureTestCasesInfrastructure1.cs
@@ -1,0 +1,18 @@
+using NotoriousTest.Infrastructures;
+
+namespace NotoriousTest.UnitTests
+{
+
+    public partial class AsyncEnvironmentUnitTests
+    {
+        public abstract class MultipleInfrastructureTestCasesInfrastructure1 : Infrastructure
+        {
+
+            public MultipleInfrastructureTestCasesInfrastructure1() : base(false)
+            {
+
+            }
+
+        }
+    }
+}