+++ NotoriousTest.Web/Applications/WebApplication.cs
@@ -3,12 +3,13 @@ using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.Configuration;
 
 using NotoriousTest.Configuration;
+using NotoriousTest.Web.Helpers;
 
 namespace NotoriousTest.Web.Applications
 {
-    public abstract class WebApplication<TEntryPoint> : WebApplicationFactory<TEntryPoint>, IConfigurable where TEntryPoint : class
+    public abstract class WebApplication<TEntryPoint> : WebApplicationFactory<TEntryPoint>, IConfigurationConsumer where TEntryPoint : class
     {
-        public Dictionary<string, string>? Configuration { get; set; }
+        public List<ConfigurationEntry<object>> ConsumedConfiguration { get; set; } = new List<ConfigurationEntry<object>>();
 
         protected override void ConfigureWebHost(IWebHostBuilder builder)
         {
@@ -16,8 +17,15 @@ namespace NotoriousTest.Web.Applications
 
             builder.ConfigureAppConfiguration((config) =>
             {
-                config.AddInMemoryCollection(Configuration);
+                Dictionary<string, string> aggregatedConfiguration = ConsumedConfiguration
+                                                .Select(ce => ce.Value.ToDictionary(ce.Key))
+                                                .SelectMany(d => d)
+                                                .ToDictionary(kvp => kvp.Key, kvp => kvp.Value);
+
+                config.AddInMemoryCollection(aggregatedConfiguration);
             });
         }
+
+
     }
 }