+++ NotoriousTest/Configuration/IConfigurationConsumer.cs
@@ -1,0 +1,7 @@
+﻿namespace NotoriousTest.Configuration
+{
+    public interface IConfigurationConsumer
+    {
+        List<ConfigurationEntry<object>> ConsumedConfiguration { get; set; }
+    }
+}